backend/requirements.txt
ifcopenshell
ifctester
langgraph
langchain
langchain-core
langchain-groq
python-dotenv
rich>=13.7.0
fastapi
uvicorn[standard]
